+++ docs/changelog.rst
@@ -13,6 +13,7 @@ Currently in development.
 - Added default permissions for other databases to user preferences
 - Added support for external links to the objects list
 - Added support for sharing with other databases to the objects list
+- Added buttons for multiple object interactions to related objects tree
 - Added filtering by object IDs and related_user_id to HTTP API objects endpoint
 - Added tree picker to location form
 - Improved .eln file support